requirements.txt
streamlit
google-generativeai
PyPDF2
langchain
langchain_community
langchain_core
